[Bronze III] Title: 컵라면 측정하기, Time: 0 ms, Memory: 1112 KB, Score: 40 point -BaekjoonHub

+++ 백준/Bronze/16479. 컵라면 측정하기/컵라면 측정하기.c
@@ -1,10 +1,10 @@
 #include<stdio.h>
 
 int main() {
-	int k, d, D;
-	scanf("%d", &k);
-	scanf("%d %d", &d, &D);
-		float m = (d - D) / 2;
-		float h = k * k - m * m;
-		printf("%f", h);
+	double k, d, D;
+	scanf("%lf", &k);
+	scanf("%lf %lf", &d, &D);
+	double m = (d - D) / 2;
+	double h = k * k - m * m;
+	printf("%g", h);
 }
